Determine annotation changes using `Equatable`
This fixes an issue in which the maps presents stale data in an annotation. The scenario happens when the contents of `annotationItems` changes, but item ids stay the same.

By comparing with `==` instead of just testing the `id` value, we ensure that changing anything within the `AnnotationItem.Element` will cause the associated annotation view to update.